Record Consumer Data at Tests End
* Move recording consumer creation timestamp to the end of the tests
